+++ nextgisweb/lookup_table/util.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+from __future__ import division, absolute_import, print_function, unicode_literals
+
+from ..i18n import trstring_factory
+
+COMP_ID = 'lookup_table'
+_ = trstring_factory(COMP_ID)